feat: turn repo in a workspace
Not sure if this has much benefit, but it saves having separate
`Cargo.lock` files for each challenge. Additionally, one can run some
commands - like `cargo test` - on all the packages in the workspace at
once, which is neat.

I think I'll have to manually add new packages to the workspace though,
that's a downside.

It's been a long time since I experimented with this change and hadn't
committed it at the time. I'll just see how it works out.